requirements.txt
apify-client>=1.0.0,<2.0.0
openai>=1.0.0
anthropic>=0.5.0
google-generativeai==0.3.0
pydantic>=2.0.0
